fix test bug
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1418500 13f79535-47bb-0310-9956-ffa450edef68

+++ solr/contrib/dataimporthandler/src/test/org/apache/solr/handler/dataimport/AbstractSqlEntityProcessorTestCase.java
@@ -72,8 +72,12 @@ public abstract class AbstractSqlEntityProcessorTestCase extends
     sportsTransformer = false;
     underlyingDataModified = false;
     
-    new File(fileLocation + File.separatorChar + fileName).delete();
-    new File(fileLocation).delete();
+    //If an Assume was tripped while setting up the test, 
+    //the file might not ever have been created...
+    if(fileLocation!=null) {
+      new File(fileLocation + File.separatorChar + fileName).delete();
+      new File(fileLocation).delete();
+    }
   }
   
   protected void logPropertiesFile() {
